refactor: Move files inside src folder
