+++ src/training/settings.py
